Fix dictation pipeline + codex-review punch list (~30 items)
Core bug fix — DictationRecorder now produces transcribable audio:
- Write native-format Float32 PCM WAV instead of letting AVAudioFile convert
  to Int16. The internal converter applies a buggy ~140x amplification that
  rail-clipped clean 0.24-peak voice to 32750/32767 Int16, destroying speech
  intelligibility for Parakeet (returned random one-word hallucinations).
- DictationAPI.swift duration_ms: Int → Double (server returns float).
- MenuBarController catch chain differentiates emptyRecording / writeFailed
  / meetingActive errors so users see meaningful UX.

Concurrency / safety (codex-review C1-C4, I1-I8, G1):
- TOCTOU race closed: stopRequested fence + ioQueue.sync drain in stop().
- Tap write errors surfaced via OSAllocatedUnfairLock<Error?> instead of
  silently logging.
- DictationError.emptyRecording thrown when frames < 50ms.
- /tmp/wispralt-last-dictation.wav debug copy gated behind #if DEBUG.
- AVAudioFile.write moved off realtime render thread onto serial ioQueue.
- AVAudioEngineConfigurationChange observer wired (catches AirPods plug-in
  mid-recording).
- Engine.prepare() before start; deinit cleanup; full error-path teardown.
- In-tap float clamp to ±0.95 with diagnostic peak tracking.

Build infrastructure:
- Package.swift linkerSettings adds @executable_path/../Frameworks rpath
  so SPM-built executable can resolve bundled Sparkle.framework. Both
  build-client.sh and build-client-local.sh inherit and verify via otool.
- build-client-local.sh: persistent self-signed identity, tar round-trip
  xattr strip (defeats macOS Tahoe provenance auto-tagging), arch detect.
- scripts/setup-local-codesign.sh: one-time setup that generates the
  cert and adds it to System trust as a code-signing root, so TCC grants
  persist across kill+relaunch. (Note: still doesn't survive rebuilds —
  Apple tightened this in Tahoe.)

Docs (G1-G3 + new findings):
- ARCHITECTURE.md DictationRecorder row rewritten for Float32 path.
- DEPLOYMENT-NOTES.md captures both AVFoundation conversion bugs and
  the TCC re-grant limitation.
- TROUBLESHOOTING.md "empty transcription / random word" entry now
  documents both historical root causes and the diagnostic float-peak
  range to verify the fix.
- API.md adds audio format flexibility section.
- SETUP-CLIENT.md and OVERVIEW.md document the new local build flow
  and the persistent-codesign one-time setup.

Tested end-to-end: server (Mac mini) p50 ≈ 181ms Parakeet inference,
client dictation roundtrip 300-900ms FN-release → text injected via AX,
multi-sentence transcription accuracy verified.

+++ client/WisprAlt/Server/DictationAPI.swift
@@ -7,7 +7,10 @@ enum DictationAPI {
     private struct TranscribeResponse: Decodable {
         let text: String
         let model_id: String
-        let duration_ms: Int
+        // Server returns this as a float (e.g. 119.94 ms), so we decode as Double.
+        // Decoding as Int would silently fail with a "data not in correct format"
+        // error because Foundation's JSONDecoder rejects float→Int coercion.
+        let duration_ms: Double
     }
 
     // MARK: - Public API
